gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
psycopg2-binary==2.9.10
redis[hiredis]==5.2.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
tiktoken==0.8.0
stripe==11.0.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-multipart==0.0.17
python-dotenv==1.0.1
